Moves `FeedStoreSpecs` to its own file - Any `FeedStore` implementation must conform to the protocol to ensure it meets all specs
