+++ EssentialFeedTests/Feed Cache/FeedStoreSpecs.swift
@@ -1,0 +1,43 @@
+//
+//  FeedStoreSpecs.swift
+//  EssentialFeedTests
+//
+//  Created by Dillon on 2/27/22.
+//
+
+import Foundation
+
+protocol FeedStoreSpecs {
+    func test_retrieve_deliversEmptyOnEmptyCache()
+    func test_retrieve_hasNoSideEffectsOnEmptyCache()
+    func test_retrieve_deliversFoundValuesOnNonEmptyCache()
+    func test_retrieve_hasNoSideEffectsOnNonEmptyCache()
+    
+    func test_insert_deliversNoErrorOnEmptyCache()
+    func test_insert_deliversNoErrorOnNonEmptyCache()
+    func test_insert_overridesPreviouslyInsertedCacheValues()
+    
+    func test_delete_deliversNoErrorOnEmptyCache()
+    func test_delete_hasNoSideEffectsOnEmptyCache()
+    func test_delete_deliversNoErrorOnNonEmptyCache()
+    func test_delete_emptiesPreviouslyInsertedCache()
+    
+    func test_feedStoreOperations_sideEffectsRunSerially()
+}
+
+protocol FailableRetrieveFeedStoreSpecs: FeedStoreSpecs {
+    func test_retrieve_deliversFailureOnFailedRetrieval()
+    func test_retrieve_hasNoSideEffectsOnFailedRetrieval()
+}
+
+protocol FailableInsertFeedStoreSpecs: FeedStoreSpecs {
+    func test_insert_deliversErrorOnFailedInsertion()
+    func test_insert_hasNoSideEffectsOnFailedInsertion()
+}
+
+protocol FailableDeleteFeedStoreSpecs: FeedStoreSpecs {
+    func test_delete_deliversErrorOnFailedDeletion()
+    func test_delete_hasNoSideEffectsOnFailedDeletion()
+}
+
+typealias FailableFeedStore = FailableRetrieveFeedStoreSpecs & FailableInsertFeedStoreSpecs & FailableDeleteFeedStoreSpecs
